[modify] restore children content

+++ app/models/concerns/ss/model/file.rb
@@ -321,6 +321,7 @@ module SS::Model::File
     backup.data = attributes
     backup.site = self.site
     backup.save
+    return unless File.exists?(path)
     trash_path = "#{Rails.root}/private/trash/#{path.sub(/.*\/(ss_files\/)/, '\\1')}"
     FileUtils.mkdir_p(File.dirname(trash_path))
     FileUtils.cp(path, trash_path)
